8pit/dwm: yet another hotfix for fullscreen mode
If the currently focused window is in fullscreen mode and another window
on the same monitor is being removed (i.e. a Terminal running a shell
with `sleep 3 && exit`) the the current window looses it's fullscreen
mode and the focus is lost.

I have no idea why the fullscreen mode is lost but by forcing it to
explicitly loose the fullscreen mode in focus() we don't lose the focus
at least.